Make config change, set in depthai, test
